Mixed two book controllers
